Add util.inherit and util.publish
Not surprisingly I was doing the same things over and over to
prototype-inherit from Node.

This is automated by the `peg/epnf` class, but we're doing a
handrolled recursive descent parse in the hope to write a proper
PEG later.

These functions do the grunt work, I'll port the classes over
soon.