Proper inverter class
We now default to Lua but can construct any number of inverters.

I will massage this further so that the module caches instances
of the inverters and builds a reasonable library of defaults, so
that `invert("python")` idempotently returns a pre-cooked python
inverter.

Desired behavior is that `invert()` returns a lua inverter,
`invert("string")` gives back language "string" and `invert({})`
hand-rolls an inverter from config values.