The Orb
I'm having a pretty good night, all things considered.

I've decided to change a letter while I still can.  Org is not a word.
Org- is a prefix, sure, but if Org were a word, it would be hanging out
with warg, and orc, and ogre.

That sounds funky and deep and I'd come sit by that fire in a heartbeat.
But that's not what the orb directory is.  It's the origin of interaction
and generation within a codex.

I like the Orb.  Orb is better.  Orb wins.

Orb.