Grymback complete
This solves a grym-specific but nonetheless vexing problem: Editing
the orb files in a way that leaves the sorcery in a bad state
was requiring me to consult the sorcery files.

Grym means never looking at the sorcery files. Worse, I had to
fix everything twice and then metacirculate.

Now I can ./grymback and away we go.