Dot compatibility
Now both dot methods accept both isnode and isNode for the
walk.

This is just a shim to get me through porting Doc to the new
Node class.

Also, dotfiles are now always created and compared, triggering
SVG creation if they change. That's the expensive command, and
the .md changing is a lagging indicator of .dot changes.