Delete orbback
In addition to being utterly hideous, in contrast to the Harambesque
`grymback`, it was janky.

I'm now just keeping a second repo around that runs a stable revision.
This works better because we're making smaller semantic changes to the
`orb` runtime, small enough that running a branch a even a couple days
old will probably reboot successfully.

That might change for awhile when we drill down on `knit` but, first
to flesh out some classes.