Aggressive interface in File brought into line
I had this great idea to control file access on the environment
level by refusing to make a file more than once.

This was not a good idea, and now File does what the others do,
cache on the basis of the string.

I need to make those table weak at some point.