Add "stage" field to version table
Really, we're renaming `edition` to `stage`.

"Edition" has a specific meaning in bridge-style semantic
versioning, and that was the original meaning of the field in
the database.

I started calling it "SNAPSHOT" and have been working up some
logic surrounding "SNAPSHOT", "CANDIDATE", and "RELEASE" stages in
the release of a published version, mainly to prevent non-RELEASE
versions of a project from being exported, at least, not without
clear labeling.