Make patch an integer
The original intention was to represent Kelvins and Knuth patch
versions as e.g. [4] or [14159..], but it makes more sense to include
them as numerals so WHERE > etc. work correctly.

I *think*, I wouldn't want coercion to turn a long knuth string into
something we can't compare againt. hmm.

It's definitely better for the common case, we'll do it this way
for now.