Use Node:inherit for Section
I'm not actually sure that continuing this work is going to prove
useful.

The Orb parser is so bad that I'm tempted to declare a flag-day
and rewrite it completely.  `espalier` is much more featureful and
powerful than it was when I wrote this parser, and I've learned a
lot about how I actually want this to work.