Refuse to make a Dir object from a non-dir path
We allow nonexistent paths, so we can make directories,
but disallow paths which happen to exist at the time of
creation.

I suppose we *could* create the file as a third return value, but
that strikes me as overly magical; such a decision can be left to
the call site. And besides, there are more filesystem types than
just 'file' and 'directory'.