Recursive italic and bold rule
This works! It does come with some caveats.

Notably, the outer capture isn't visible in the inner capture, so
that this:

`***bold /italic *bold* italic/ bold***`

is marked up with two bold groups, while

`***bold *still bold* bold***` is just one group.

This is... okay. The real rule will have backslash-escaping for
the serious edge cases, and most `*` or `/` that aren't intended
as markup won't be matched pairs.