Match by name and number of !,/ in code block
Once again, I have to make an anonymous group in order to have
back references work.

This is actually ok, because of the inner-parsing ability. It does
feel a bit like a workaround however.

Still, it syntactically validates, which is pretty neat.  It means
that, when I go to parse a code block, I don't have to check that
the bookends match; if they didn't, it wouldn't be a code block.