Hide block rule
I think it will be more tractable to iterate sub-sectional parts
directly, rather than to select() on "block", especially since
that will bring in any block rule in subsidiary sections.

Now, there are times when that behavior is useful, but I think
those will only benefit from being specific.

It's easy to remove the backticks if that proves more useful.