Use >= for chunking lines
Just added this feature for comparing matches based on length.

I will probably refine these rules so that for a line that's
not indented, it will require strictly greater than.  This
gets around an edge case:

```
- list item

  A new paragraph, with the
spacing that is sometimes more attractive
```

We don't *require* double-spacing after a list marker, but we do
encourage it if there's not a check box, so this should be okay.