Begin to convert grimoire.orb to Orb format
It's time to be working on the documentation in concert with the
new parser.

Right now, I have two docs, grimoire.orb and orb.orb, which
overlap but represent different eras of thought.  Long-term, the
only thing for it is to consolidate into one document.

Medium-term, grimoire.orb was originally an org document and
retains traces of that history.  Converting it and marking it up
with notes and easy corrections is an important step.