Add tiebreaking indentation for lists
I'm not sure how I feel about this.

It has to be one way or another: either an indented list loses the
occasional element, or a list accidentally ingests a line from an
indented paragraph.

I *think* the latter is worse, but I consider indenting the start
of every paragraph to be bad style, but also, good style for a
list is two spaces between the key or structure mark, and the
content.

I'll leave it this way for now, and see what happens.

Come to think of it, it's possible to prevent this by making a
block-sep mandatory after a list.  Hmm.