Add table rule to Doc grammar
Named captures behave badly with the rules we've established.

I think this can be fixed, maybe, by using Cp() and a sub action
rule, which would then be added to Elpatt to make it easier to
use.

For now, I'm leaning toward making a _c version of each named
capture rule in fragments, and using that as the body of the
non- _c version, so we'd have

handle <- handle_c
`handle_c` <- "@" symbol

Because we're really trying to define each rule in one place.