Remove Raw printing with strLine metamethod
Hiding intermediate results is useful for distinctions which are
syntactically useful but not gramattically relevant.

A Doc Node should be a concrete syntax tree, one in which every
byte of the source file is represented as a leaf node.

So even whitespace will be present in the final AST.  But this
will make the REPL output difficult to use, so we simply return
the empty string for nodes we don't need to look at.