Working list-line final cookies
The problem is in the PEG parser, this:

rule <- 1+ -1

turns into

P(1) - P(1)

not

P(1) * -P(1)

The workaround is to group the -1, but I do want to fix the parser.

I think the two ways of expressing negative lookahead aren't pulling
their weight at this point, and I should go back to - only being
valid in the middle of rule names and before a number.