Refine prose rules
This is going to get truly gnarly when I add logic to parse words
and punctuation inside of structure blocks.

I'm not even sure how to accomplish it, to be quite honest. Ah
well, a task for another time.

I might even just do a post-pass that looks for gaps and shotguns
them into Nodes, it's conceptually cleaner than other approaches
I have in mind.

There are a lot of contortions necessary to keep back captures in
play :/