Memoized Twig:select(pred)
I want to do as few passes as possible through a Doc, so we
memoize a given select() statement the first time we use it.

I think this will prove worthwhile once we start to add literate
programming, although we'll see.

My thinking was that we'll do a lot of looking through #tags and
the like. I dunno, it occurred to me in a hypogogic state, I've
added it, and it doesn't change the semantics of selection provided
that the fundamental structure of the tree doesn't change.

I think that any changes big enough to modify the tree structure
would result in a reparsing of a new string, but we'll see.