Specialize knitter collection by code_type
This isn't actually perfect, because it doesn't test the
predicates, but it's decent for now.

Whether or not to do this is a bit of a judgement call, and I
might walk it back.  My thinking is that we'll eventually have a
knitter for every conceivable programming language.

Eh. We'll revisit this when the time comes, I'm sure.