Strip leading whitespace from 'compiled' strings.
This gives a ration of 1.169 between strings and bytecode over the
whole Orb suite.

That's not really worth the hassle of bytecode not being a stable
interface; I'll probably keep the stringified form.

This commit includes the extra ratio information, which I'll use
across a few projects out of curiousity.