Remove header-text category
There's no particular reason not to include the leading whitespace
in the head-line.

We intend to parse inside the head-line anyway, and it may as well
have the space in it.

Technically we don't even need to include the " " in that rule,
since we've already matched it in the main Doc grammar.