Remove 'scrolls' level of indirection from Skein
I vaguely remember why I thought that was a good idea, but in
fact, there's no need for the `knitted` map to hold anything other
than the various scrolls knitted from the Doc.

I was going to put completion closures in a separate table, but
there's no need, they can go directly on the Lume where they will
be useful.