Handle brackets in handles "correctly"
This performs matching on brackets, so that @name[subname] will
parse, but @name]subname[ and @name[subname will just be @name.

It doesn't allow gaps inside the brackets, which I think is
correct.

That said, I should give quite a bit more thought to what makes a
name valid.

This fixes a problem I was having with helm, so it will stand for
now.