Expand grammar for link refs
This breaks things up into appropriate pieces.

It doesn't go as far as it could, in terms of parsing inside the
distinct project/doc/fragment bits.

And it needs a bit of massaging; I'm pretty sure I want different
names for handle-refs which don't have a col in them, because those
are semantically more like a fragment.

It'll get us where we want to go, though.