Switch parsers to native PEG blocks
It's a satisfying feeling to be writing a genuinely polyglot
program.

The 'inner language' effect of embedding another grammar inside
strings is an artifact of our programming paradigm.

Ridding myself of it is a powerful demonstration that literate
programming can make things better.

Once I add tranclusion (and I'm close!), these native blocks can
be reused elsewhere.  With SQL in particular I expect this to be a
force multiplier.