Strip all old code from orb:orb
What's left is just a container for the Lume, and we could replace
the whole thing with `return require "orb:lume/lume" and call it
a day, after modifying pylon.

But we're going to leave Orb as a library with one constructor in
it, because I expect I'll just have to turn it back into that at
some point if we don't.

If we go long enough without having other entrances to the system,
I'll just replace it.