Kinda-sorta fill in the scroll
I think the next step is to stop using Phrases, which aren't
actually helping me once we switch to an all-scroll method.

They also seem to be buggy when I start doing weird things to them,
which is alarming given how strongly they underly what we've been
doing so far.