Fixed enough for now
This is a large diff, because our section logic was off in scroll,
which was leaving off the final newline on what might have been
every knit.

I didn't notice, because it *looked* like there was a final newline,
because it was `return whatever\n`, but there should be a blank
line for the `#/lua` block, and a newline after that.

The markdown weaver is subtly broken in all sorts of places, but
that's not related to the scroll-for-weave push, at least, not
directly, so I'm closing the case on this one.