Key now one non-gap long
I had a weird bit in the weave where a colon in the middle of a
multi-line list entry was causing parse errors, so I've decided to
restrict keys to one not-gap string with whitespace on both ends.

Keys will probably need to be more complex than that, but it will
do for now.