Remove sha2, use util:sha
We need the sha function in at least one other place.

This commit actually changes the nature of the sha, using the
literal string for anything shorter than the sha itself.  This
may be expected to have minimal impact on orb; it may break
uniqueness for some extremely short modules, but not in a way which
will cause problems, especially not on a reasonable time horizon.