Credible list tagging
Took a break to write an annotator, which is a little note-taking
gizmo that I can attach to a Skein or whatever else, and use to
smuggle notes into helm.

This really sped up development, it's a bit awkward to use
right now but as a proof of concept for how helm status printing
might work it's pretty great.

Anyway. List tagging works for my test case, sections next, then
I'll tackle the cling rule.