Use tags in #asLua predicator
This means that #AsLua behaves the same as #asLua, which is a
guarantee we make in the documentation about how tags work.

It also means we pass the entire skein to the predicate functions,
which feels kind of reckless? but it might come in handy?

Like I could picture something from the manifest proving to be
important for predicated knitting. Idk.