Manual iteration of subsections
I have zero understanding of why running :select "section" on the
section returns only the section itself and not any of the
subsections. It really feels like the sort of thing I should be
able to answer.

But this does the job, and actually examines a lot less nodes in
the process, so I'm just going to leave it alone for now.