Separate out action of walk loop into own function
This is with an eye toward adding one walking step which takes
various actions with respect to specific kinds of Node.

Not sure this will prove possible, for instance, some of what I
might want to do to codeblocks will require the tagging step to
be completed, but I do want the option.