Refine hashtag rule
Handles need to be pretty free-form, and we may end up with several
rules depending on where the handle can be found (that is, a ref
might be syntactically distinct from a name), although I don't want
to do that.

Hashtags aren't intended as a user-extensible system, Orb will
provide the canonical set of hashtags.  The user can put in a
hashtag which Orb doesn't know how to use, but there's no reason to
allow for hashtags which aren't symbol-like.

This rule allows hashtags to be namespaced, but also parses
`#Todo:` as the hashtag `#Todo` followed by a colon, while
`#Todo:done` is all one hashtag.