Add trailing hashtags to link grammar
This lets us add attributes or otherwise instruct the engine to
treat a link special, examples are #img or #rel-nofollow.

To make it obvious that it isn't an obelus, we don't provide a
short form for these, but [[][url]#hashtag] will parse just fine
and the engine will interpret that as the url being the text of
the link.