Cleaner note slug
I'm probably going to leave it this way.

It's tempting to make the note syntax itself into `{fn:1}` without
the square brackets, but I think that a) it's too easy to convert
random uses of curly brackets into spurious notes and b) the square
brackets make the note look more 'link-like', especially in the
wrapped form.