Begin global manifest read in lume
This will give us the ability to set a Manifest for absolutely
everything, which will be pretty useful.

I expect we'll also want to check parent directories, so that
someone can have a project root with specific overrides, and the
order will go: global, parent(s), project-local, and any #manifest
block in a Doc.